# frontend/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "AI Learning Companion API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/conversations/paste" : {
      "post" : {
        "operationId" : "paste",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IntakeResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "conversation-input-controller" ]
      }
    },
    "/api/conversations/share-link" : {
      "post" : {
        "operationId" : "shareLink",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ShareLinkRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IntakeResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "conversation-input-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "IntakeResponse" : {
        "type" : "object",
        "properties" : {
          "conversationId" : {
            "type" : "string"
          },
          "inputPath" : {
            "type" : "string"
          },
          "learningSessionId" : {
            "type" : [ "integer", "null" ],
            "format" : "int64"
          },
          "reviewUnitCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "userTurnCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "warnings" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "conversationId", "inputPath", "learningSessionId", "reviewUnitCount", "userTurnCount", "warnings" ]
      },
...
      "PasteRequest" : {
        "type" : "object",
        "properties" : {
          "text" : {
            "type" : "string"
          }
        },
        "required" : [ "text" ]
      },
...
      "ShareLinkRequest" : {
        "type" : "object",
        "properties" : {
          "url" : {
            "type" : "string"
          }
        },
        "required" : [ "url" ]
      },
...
    }
  }
}